Update unit tests + Replace entirely old VEP class
